+++ src/main/java/com/maxmind/minfraud/request/Account.java
@@ -2,7 +2,11 @@ package com.maxmind.minfraud.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.maxmind.minfraud.AbstractModel;
-import org.apache.commons.codec.digest.DigestUtils;
+
+import java.math.BigInteger;
+import java.nio.charset.Charset;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
 
 /**
  * Account related data for the minFraud request
@@ -46,8 +50,15 @@ public final class Account extends AbstractModel {
          * @return The builder object.
          */
         public Account.Builder username(String username) {
-            this.usernameMd5 = DigestUtils.md5Hex(username);
-            return this;
+            try {
+                MessageDigest d = MessageDigest.getInstance("MD5");
+                d.update(username.getBytes(Charset.forName("UTF8")));
+                BigInteger i = new BigInteger(1, d.digest());
+                this.usernameMd5 = String.format("%032x", i);
+                return this;
+            } catch (NoSuchAlgorithmException e) {
+                throw new RuntimeException("No MD5 algorithm for MessageDigest!", e);
+            }
         }
 
         /**